Update 2012-09-12-http-response-codes.md
Added the word 'been' on rule #78. Was: "For example, imagine a Web address giving the weather forecast for Taipei, and there has lately a very strong typhoon.". Is now: "For example, imagine a Web address giving the weather forecast for Taipei, and there has lately been a very strong typhoon."